Move NIHStatus object
