+++ src/main/scala/org/broadinstitute/dsde/firecloud/service/NihService.scala
@@ -5,14 +5,47 @@ import akka.pattern._
 import com.typesafe.scalalogging.slf4j.LazyLogging
 import org.broadinstitute.dsde.firecloud.Application
 import org.broadinstitute.dsde.firecloud.dataaccess.{RawlsDAO, ThurloeDAO}
-import org.broadinstitute.dsde.firecloud.model.{NIHStatus, UserInfo}
+import org.broadinstitute.dsde.firecloud.model.{Profile, UserInfo}
 import org.broadinstitute.dsde.firecloud.service.NihService.GetStatus
 import org.broadinstitute.dsde.firecloud.service.PerRequest.{PerRequestMessage, RequestComplete}
+import org.broadinstitute.dsde.firecloud.utils.DateUtils
 import spray.http._
 import spray.httpx.SprayJsonSupport._
+import spray.json.DefaultJsonProtocol._
 
 import scala.concurrent.{ExecutionContext, Future}
 
+
+case class NIHStatus(
+  loginRequired: Boolean,
+  linkedNihUsername: Option[String] = None,
+  isDbgapAuthorized: Option[Boolean] = None,
+  lastLinkTime: Option[Long] = None,
+  linkExpireTime: Option[Long] = None,
+  descriptionSinceLastLink: Option[String] = None,
+  descriptionUntilExpires: Option[String] = None
+)
+
+object NIHStatus {
+
+  implicit val impNihStatus = jsonFormat7(NIHStatus.apply)
+
+  def apply(profile: Profile): NIHStatus = {
+    apply(profile, profile.isDbgapAuthorized)
+  }
+
+  def apply(profile: Profile, isDbGapAuthorized: Option[Boolean]): NIHStatus = {
+    val linkExpireSeconds = profile.linkExpireTime.getOrElse(0L)
+    val howSoonExpire = DateUtils.secondsSince(linkExpireSeconds)
+    new NIHStatus(
+      loginRequired = howSoonExpire >= 0,
+      profile.linkedNihUsername,
+      isDbGapAuthorized,
+      profile.lastLinkTime,
+      profile.linkExpireTime
+    )
+  }
+}
 
 object NihService {
   sealed trait ServiceMessage
